Drop ifupdown2
It only served to semplify VRRP configuration, which is a niche feature, and it
is perfectly possible to do with iproute2. ifupdown2 gave me a lot of headaches
with networking.service being disabled, ifup@.service reporting that ifup was
already running, and there is some buzz about such issues, e.g.

    https://forum.proxmox.com/threads/ifupdown2-no-network-enable-on-reboot-if-purged-ifupdown.74697/
    https://github.com/CumulusNetworks/ifupdown2/issues/177
    https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=988050

I feel that for the little use it has, it is not worth the pain of debugging and
solving all the issues.